+++ firebase.go
@@ -18,11 +18,12 @@
 package firebase
 
 import (
-	"context"
 	"encoding/json"
 	"errors"
 	"io/ioutil"
 	"os"
+
+	"golang.org/x/net/context"
 
 	"cloud.google.com/go/firestore"
 
@@ -41,7 +42,7 @@ import (
 var defaultAuthOverrides = make(map[string]interface{})
 
 // Version of the Firebase Go Admin SDK.
-const Version = "2.6.0"
+const Version = "2.6.1"
 
 // firebaseEnvName is the name of the environment variable with the Config.
 const firebaseEnvName = "FIREBASE_CONFIG"